refactor(rsc): separate app route ownership
