+++ packages/plugin-rsc/examples/action-reachability/src/app/a/action-indirect.ts
