+++ packages/plugin-rsc/examples/action-reachability/src/app/a/action.tsx
@@ -1,6 +1,6 @@
 'use server'
 
-import { getActionContext } from '../../framework/action-context.ts'
+import { getActionContext } from '../action-context.ts'
 
 export async function actionA() {
   return `ACTION_A_OK:${getActionContext().route}`
