+++ packages/plugin-rsc/examples/action-reachability/src/app/a/middleware.ts
@@ -1,5 +1,5 @@
-import { runWithActionContext } from '../../framework/action-context.ts'
 import type { RouteMiddleware } from '../../framework/middleware.ts'
+import { runWithActionContext } from '../action-context.ts'
 
 export const middleware: RouteMiddleware = (request, next) =>
   runWithActionContext({ request, route: '/a' }, next)
