+++ packages/plugin-rsc/examples/action-reachability/src/app/b/action.tsx
@@ -1,6 +1,6 @@
 'use server'
 
-import { getActionContext } from '../../framework/action-context.ts'
+import { getActionContext } from '../action-context.ts'
 
 export async function actionB() {
   return `ACTION_B_OK:${getActionContext().route}`
