+++ packages/plugin-rsc/examples/action-reachability/src/app/b/page.tsx
